content/xslt/tests/mochitest/test_bug551412.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=551412
-->
<head>
  <title>Test for Bug 551412</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=551412">Mozilla Bug 551412</a>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 551412 **/

  var processor = new XSLTProcessor();
  var style =
    '<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" ' +
                    'xmlns:exsl="http://exslt.org/common" ' +
                    'version="1.0">' +
      '<xsl:output method="html"/>' +
      '<xsl:variable name="rtf">1 <b>2</b> 3</xsl:variable>' +
      '<xsl:template match="/">' +
        '<xsl:copy-of select="exsl:node-set($rtf)"/>' +
      '</xsl:template>' +
    '</xsl:stylesheet>';
  var styleDoc = new DOMParser().parseFromString (style, "text/xml");

  var data =
    '<root/>';
  var originalDoc = new DOMParser().parseFromString(data, "text/xml");

  processor.importStylesheet(styleDoc);

  var fragment = processor.transformToFragment(originalDoc, document);
  var content = document.getElementById("content");
  content.appendChild(fragment);
  is(content.innerHTML, "1 <b>2</b> 3",
     "Result of transform should be '1 <b>2</b> 3'");

</script>
</pre>
</body>
</html>
